Separacion de posts en nuevo fragment
El fragment que antes se llamaba HistoriasFragments ahora es CategoriasFragment y solo muestra las categorias. Un nuevo fragment se encarga de solo mostrar los posts cuando se toca una categoria. Hay un problema donde no se pueden ver los posts todavia, pero si se reciben bien de firebase
