+++ app/src/main/java/com/wolkorp/petrescue/models/Category.kt
@@ -1,3 +1,3 @@
 package com.wolkorp.petrescue.models
 
-data class Category(val imageURL: String, val categoryName: String)
+data class Category(val imageURL: String, val categoryName: String, val categoryDescription: String)
