+++ app/src/main/res/layout/fragment_categorias.xml
@@ -1,0 +1,101 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".fragments.CategoriasFragment">
+
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/profile_picture"
+        android:layout_width="46dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="16dp"
+        android:elevation="8dp"
+        app:cardCornerRadius="20dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:scaleType="centerCrop"
+            android:src="@drawable/foto_perfil" />
+
+    </androidx.cardview.widget.CardView>
+
+
+    <androidx.appcompat.widget.AppCompatTextView
+        android:id="@+id/mensaje_bienvenida"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
+        android:text="Hola,"
+        android:textColor="#000000"
+        android:textSize="18sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="@+id/profile_picture"
+        app:layout_constraintEnd_toStartOf="@+id/profile_picture"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/profile_picture" />
+
+
+    <androidx.appcompat.widget.AppCompatTextView
+        android:id="@+id/frase_entrada"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
+        android:text="¿Como quieres ayudar"
+        android:textColor="#000000"
+        android:textSize="30sp"
+        android:textStyle="bold"
+        app:layout_constraintTop_toBottomOf="@+id/profile_picture" />
+
+
+    <androidx.appcompat.widget.AppCompatTextView
+        android:id="@+id/segunda_frase_entrada"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
+        android:layout_marginEnd="20dp"
+        android:text="hoy dia?"
+        android:textColor="#757575"
+        android:textSize="30sp"
+        app:layout_constraintTop_toBottomOf="@id/frase_entrada" />
+
+
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/categoriesViewPager"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginTop="30dp"
+        android:layout_marginBottom="30dp"
+        android:paddingStart="50dp"
+        android:paddingEnd="50dp"
+        app:layout_constraintTop_toBottomOf="@id/segunda_frase_entrada"
+        app:layout_constraintBottom_toBottomOf="parent" />
+
+
+ <!--commentario por ahora
+
+      <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/categoriesViewPager"
+            android:layout_width="411dp"
+            android:layout_height="132dp"
+            android:layout_marginTop="12dp"
+            android:paddingStart="40dp"
+            android:paddingEnd="40dp"
+            app:layout_constraintTop_toBottomOf="@+id/segunda_frase_entrada"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+
+-->
+
+
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
