+++ app/src/main/res/layout/fragment_historias.xml
@@ -8,106 +8,22 @@
     tools:context=".fragments.HistoriasFragment">
 
 
-    <androidx.cardview.widget.CardView
-        android:id="@+id/profile_picture"
-        android:layout_width="46dp"
-        android:layout_height="40dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="16dp"
-        android:elevation="8dp"
-        app:cardCornerRadius="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:scaleType="centerCrop"
-            android:src="@drawable/foto_perfil" />
-
-    </androidx.cardview.widget.CardView>
-
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/mensaje_bienvenida"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:text="Hola,"
-        android:textColor="#000000"
-        android:textSize="18sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@+id/profile_picture"
-        app:layout_constraintEnd_toStartOf="@+id/profile_picture"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/profile_picture" />
-
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/frase_entrada"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:text="Frase de bienvenida"
-        android:textColor="#000000"
-        android:textSize="30sp"
-        android:textStyle="bold"
-        app:layout_constraintTop_toBottomOf="@+id/profile_picture" />
-
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:id="@+id/segunda_frase_entrada"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginEnd="20dp"
-        android:text="Categorias"
-        android:textColor="#757575"
-        android:textSize="30sp"
-        app:layout_constraintTop_toBottomOf="@id/frase_entrada" />
-
-<!--commentario por ahora
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/categoriesViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginBottom="150dp"
-        android:paddingStart="40dp"
-        android:paddingEnd="40dp"
-        app:layout_constraintTop_toBottomOf="@id/segunda_frase_entrada"
-        app:layout_constraintBottom_toBottomOf="parent" />
--->
-
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/categoriesViewPager"
-        android:layout_width="411dp"
-        android:layout_height="132dp"
-        android:layout_marginTop="12dp"
-        android:paddingStart="40dp"
-        android:paddingEnd="40dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/segunda_frase_entrada" />
-
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rec_posts"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/categoriesViewPager"
+        android:layout_height="500dp"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
+        app:layout_constraintTop_toBottomOf="@+id/btnAddPost" />
 
 
     <Button
         android:id="@+id/btnAddPost"
         android:layout_width="52dp"
         android:layout_height="49dp"
+        android:layout_marginStart="340dp"
+        android:layout_marginTop="4dp"
         android:background="#00D3C3C3"
         android:foreground="@mipmap/ic_launcher"
         app:layout_constraintBottom_toTopOf="@+id/categoriesViewPager"
